+++ tools/gen_client.py
@@ -70,6 +70,7 @@ class {{ uc_async }}Client(ClientAppMixin, {{ uc_async }}ClientMixin):
             raise ValueError("Required setting 'api_url' not configured")
         self._transport = _transport
         self._debug = _debug
+        self._init_client_runtime()
 
     @property
     def config(self) -> ClientConfig:
@@ -128,7 +129,6 @@ def generate_api_code(schema: OASchema, models: set[str], is_async: bool) -> str
                 path, method_name, method, models, is_async=is_async
             )
             functions.append(function_code)
-    functions.append(generate_close_method_code(is_async))
     return "\n\n".join(functions)
 
 
@@ -227,23 +227,6 @@ def generate_function_code(
     )
 
 
-def generate_close_method_code(is_async: bool) -> str:
-    if is_async:
-        return (
-            f"{C_TAB}async def close(self):\n"
-            f'{C_TAB}{C_TAB}"""Close this client."""\n'
-            f"{C_TAB}{C_TAB}if self._transport is not None:"
-            f"{C_TAB}{C_TAB}{C_TAB}await self._transport.async_close()"
-        )
-    else:
-        return (
-            f"{C_TAB}def close(self):\n"
-            f'{C_TAB}{C_TAB}"""Close this client."""\n'
-            f"{C_TAB}{C_TAB}if self._transport is not None:"
-            f"{C_TAB}{C_TAB}{C_TAB}self._transport.close()"
-        )
-
-
 def generate_function_doc(method: OAMethod) -> str:
     doc_lines = method.description.split("\n") if method.description else []
 
@@ -316,7 +299,9 @@ def generate_function_doc(method: OAMethod) -> str:
     wrapped_doc_lines = []
     for line in doc_lines:
         if line and not line.startswith("|"):
-            for subline in textwrap.wrap(line, 72):
+            for subline in textwrap.wrap(
+                line, 72, break_long_words=False, break_on_hyphens=False
+            ):
                 # TODO: prefix `subline` 1..N with leading whitespaces from `line`
                 wrapped_doc_lines.append(subline)
         else: